+++ src/main/scala/de/m7w3/signal/Main.scala
@@ -6,12 +6,10 @@ import java.security.Security
 
 import org.whispersystems.libsignal.logging.SignalProtocolLoggerProvider
 
-import scala.reflect.runtime.universe._
 import scalafx.Includes._
 import scalafx.application.JFXApp
 import scalafx.application.JFXApp.PrimaryStage
-import scalafx.scene.{Parent, Scene}
-import scalafxml.core.{DependenciesByType, FXMLView}
+import scalafx.scene.Scene
 
 object App {
   val NAME = "signal-desktop"
@@ -29,10 +27,6 @@ object Main extends JFXApp {
   signalDesktopConfig.foreach { config =>
     val appContext = ApplicationContext(config)
 
-    val dependencies = Map[Type, Any](
-      typeOf[ApplicationContext] -> appContext
-    )
-
     val root = if (appContext.profileDirExists && appContext.profileIsInitialized) {
       UnlockDB.load(appContext)
     } else {
@@ -46,12 +40,6 @@ object Main extends JFXApp {
       title = "Welcome"
       scene = new Scene(root)
     }
-  }
-
-  def loadFXML(resourceUri: String, dependencies: Map[Type, Any]): Parent = {
-    val fxmlUri = getClass.getResource(resourceUri)
-    require(fxmlUri != null, s"$resourceUri not found")
-    FXMLView(fxmlUri, new DependenciesByType(dependencies))
   }
 
   override def stopApp(): Unit = {
