+++ src/main/scala/de/m7w3/signal/controller/ChatsListController.scala
@@ -5,14 +5,18 @@ import scalafx.event.ActionEvent
 import scalafx.scene.control.{Button, ButtonBar}
 import scalafx.scene.layout.AnchorPane
 import scalafxml.core.macros.{nested, sfxml}
+import scalafx.scene.Parent
+import de.m7w3.signal.ApplicationContext
+import scala.reflect.runtime.universe.{Type, typeOf}
+import scalafxml.core.{DependenciesByType, FXMLView}
 
 @sfxml
 class ChatsListController(
-    val chatsListButtonBar: ButtonBar,
-    val newChatBtn: Button,
-    @nested[EditMenuController] editMenuController: MenuController,
-    @nested[FileMenuController] fileMenuController: MenuController,
-    @nested[HelpMenuController] helpMenuController: MenuController) {
+  val chatsListButtonBar: ButtonBar,
+  val newChatBtn: Button,
+  @nested[EditMenuController] editMenuController: MenuController,
+  @nested[FileMenuController] fileMenuController: MenuController,
+  @nested[HelpMenuController] helpMenuController: MenuController) {
 
   println(s"editMenuController initialized $editMenuController")
 
@@ -30,5 +34,17 @@ class ChatsListController(
     val pane = new AnchorPane()
     pane.children.add(btn)
     scene.setRoot(pane)
+  }
+}
+
+object ChatsList{
+  def load(context: ApplicationContext): Parent = {
+    val dependencies = Map[Type, Any](
+      typeOf[ApplicationContext] -> context
+    )
+    val resourceUri = "/de/m7w3/signal/recent_chats_list.fxml"
+    val fxmlUri = getClass.getResource(resourceUri)
+    require(fxmlUri != null, s"$resourceUri not found")
+    FXMLView(fxmlUri, new DependenciesByType(dependencies))
   }
 }
